refactor dropdowns to reuseable components
